backend/app/services/requirements.txt
torch
transformers
numpy<2.0
pytesseract
pillow
tools
fitz
PyMuPDF
opencv